Conflicts:
	Build/Build.bat
	Build/Installer/Installer.nsi
	Library/Library.vcxproj
	Plugins/PluginNowPlaying/taglib/asf/asffile.cpp
	Plugins/PluginNowPlaying/taglib/flac/flacfile.cpp
	Plugins/PluginNowPlaying/taglib/mpeg/mpegproperties.cpp
	Plugins/PluginNowPlaying/taglib/ogg/vorbis/vorbisproperties.cpp
	Plugins/PluginNowPlaying/taglib/toolkit/tbytevector.cpp
	Plugins/PluginNowPlaying/taglib/toolkit/tfile.cpp
	Plugins/PluginNowPlaying/taglib/toolkit/tstring.cpp
	Rainmeter.sln
	Version.cs
	Version.h

+++ Plugins/PluginNowPlaying/NowPlaying.cpp
@@ -79,7 +79,7 @@ PLUGIN_EXPORT void Reload(void* data, void* rm, double* maxValue)
 		{
 			// PlayerName starts with [ so use referenced section
 			++str;
-			int len = wcslen(str);
+			int len = static_cast<int>(wcslen(str));
 			if (len > 0 && str[len - 1] == L']')
 			{
 				--len;
